backend/requirements-docker.txt
# Chi can cho script.py - loai bo anaconda/conda/comfyui (khong co tren PyPI)
Flask>=3.1.0
neo4j>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
llmware>=0.4.0
chromadb>=0.4.0
Jinja2>=3.1.0
sentence-transformers>=2.0.0
